feature: implement RelayCommand and RelayCommandT
