+++ Stopify.Presentation/Utilities/Commands/Base/RelayCommand.cs
@@ -1,0 +1,9 @@
+﻿namespace Stopify.Presentation.Utilities.Commands.Base;
+
+public class RelayCommand : RelayCommand<object>
+{
+    public RelayCommand(Action execute, Func<bool> canExecute = null)
+        : base(_ => execute(), _ => canExecute?.Invoke() ?? true)
+    {
+    }
+}
