--- Stopify.Presentation/Utilities/Commands/RelayCommand.cs
@@ -1,23 +1,0 @@
-﻿using System.Windows.Input;
-
-namespace Stopify.Presentation.Utilities.Commands;
-
-public class RelayCommand : ICommand
-{
-    private readonly Action _execute;
-    private readonly Func<bool> _canExecute;
-
-    public RelayCommand(Action execute, Func<bool> canExecute = null)
-    {
-        _execute = execute ?? throw new ArgumentNullException(nameof(execute));
-        _canExecute = canExecute;
-    }
-
-    public event EventHandler CanExecuteChanged;
-
-    public bool CanExecute(object parameter) => _canExecute == null || _canExecute();
-
-    public void Execute(object parameter) => _execute();
-
-    public void RaiseCanExecuteChanged() => CanExecuteChanged?.Invoke(this, EventArgs.Empty);
-}
